Extract badger iterators to own file
